requirements.txt
boto3
Pillow
tensorflow
flask
requests
numpy